fix login adn refactor stack file structure
