+++ emphasis_education_app/App.tsx
@@ -9,9 +9,9 @@
  */
 
 import * as React from 'react';
-import { View, Text, ActivityIndicator } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
+// import { View, Text, ActivityIndicator } from 'react-native';
+// import { NavigationContainer } from '@react-navigation/native';
+// import { createStackNavigator } from '@react-navigation/stack';
 import { ApolloClient } from 'apollo-client';
 import { InMemoryCache, NormalizedCacheObject } from 'apollo-cache-inmemory';
 import { HttpLink } from 'apollo-link-http';
@@ -20,27 +20,28 @@ import { split } from 'apollo-link';
 import { ApolloProvider, useQuery, useMutation } from '@apollo/react-hooks';
 import { getMainDefinition } from 'apollo-utilities';
 import { ThemeProvider } from 'styled-components';
-import AsyncStorage from '@react-native-community/async-storage';
+// import AsyncStorage from '@react-native-community/async-storage';
 
-import Login from './src/components/Login';
-import Chat from './src/components/Chat';
-import HomePage from './src/components/HomePage';
-import ChatPicker from './src/components/ChatPicker';
-import Profile from './src/components/LiftedProfile';
-import CreateUserContain from './src/components/CreateUserContainer';
-import ConfirmationScreen from './src/components/ConfirmationScreen';
-import Search from './src/components/AdminPage/LiftedSearch';
-import AdminPage from './src/components/AdminPage';
-import CreateChat from './src/components/CreateChat';
-import Settings from './src/components/Settings';
+// import Login from './src/components/Login';
+// import Chat from './src/components/Chat';
+// import HomePage from './src/components/HomePage';
+// import ChatPicker from './src/components/ChatPicker';
+// import Profile from './src/components/LiftedProfile';
+// import CreateUserContain from './src/components/CreateUserContainer';
+// import ConfirmationScreen from './src/components/ConfirmationScreen';
+// import Search from './src/components/AdminPage/LiftedSearch';
+// import AdminPage from './src/components/AdminPage';
+// import CreateChat from './src/components/CreateChat';
+// import Settings from './src/components/Settings';
 
 import { theme } from './src/theme';
 import context, {EmptyUser, AuthContext} from './src/components/Context/Context';
 import { IUser, ILoginPayload } from './src/types';
 
-import { LOGIN_TOKEN } from './src/constant';
+// import { LOGIN_TOKEN } from './src/constant';
+// import { LOGIN } from './src/queries/Login';
 
-import { LOGIN } from './src/queries/Login';
+import StackNavigation from './StackNavigation';
 
 const cache = new InMemoryCache();
 const httplink = new HttpLink({
@@ -73,143 +74,143 @@ const client: ApolloClient<NormalizedCacheObject> = new ApolloClient({
 });
 
 
-const AuthStackNav = createStackNavigator();
-const AuthStack = () => (
-  <AuthStackNav.Navigator initialRouteName={'Login'}>
-    <AuthStackNav.Screen
-      name='Login'
-      component={Login}
-    />
-    <AppStackNav.Screen
-      name="CreateUserContain"
-      component={CreateUserContain}
-      options={{ title: 'Create User' }}
-    />
-    <AppStackNav.Screen
-      name="ConfirmationScreen"
-      component={ConfirmationScreen}
-      options={{ title: '' }}
-    />
-  </AuthStackNav.Navigator>
-)
+// const AuthStackNav = createStackNavigator();
+// const AuthStack = () => (
+//   <AuthStackNav.Navigator initialRouteName={'Login'}>
+//     <AuthStackNav.Screen
+//       name='Login'
+//       component={Login}
+//     />
+//     <AppStackNav.Screen
+//       name="CreateUserContain"
+//       component={CreateUserContain}
+//       options={{ title: 'Create User' }}
+//     />
+//     <AppStackNav.Screen
+//       name="ConfirmationScreen"
+//       component={ConfirmationScreen}
+//       options={{ title: '' }}
+//     />
+//   </AuthStackNav.Navigator>
+// )
 
 
-const AppStackNav = createStackNavigator();
+// const AppStackNav = createStackNavigator();
 
-const AppStack = () => {
-  return (
-    <AppStackNav.Navigator initialRouteName={'Home'}>
-      <AppStackNav.Screen
-        name="Home"
-        component={HomePage}
-        options={{
-          title: '' ,
-          // we chouldnt really be able to go back to the login page?
-          // headerLeft: () => null
-        }}
-      />
-      <AppStackNav.Screen
-        name="Search"
-        component={Search}
-        options={{ title: '' }}
-      />
-      <AppStackNav.Screen
-        name="Chat"
-        component={Chat}
-        options={{ title: 'Test Subject' }}
-      />
-      <AppStackNav.Screen
-        name="ChatPicker"
-        component={ChatPicker}
-        options={{
-          title: 'My Chats',
-        }}
-      />
-      <AppStackNav.Screen
-        name="MyProfile"
-        component={Profile}
-        options={{ title: 'My Profile' }}
-      />
-      <AppStackNav.Screen
-        name="AdminPage"
-        component={AdminPage}
-        options={{ title: 'Admin Page' }}
-      />
-      <AppStackNav.Screen
-        name='CreateChat'
-        component={CreateChat}
-        options={{ title: 'New Chat'}}
-      />
-      <AppStackNav.Screen
-        name='Settings'
-        component={Settings}
-        options={{ title: 'Settings'}}
-      />
-    </AppStackNav.Navigator>
-  )
-}
-const RootStackNav = createStackNavigator();
-const RootStack = ({ userToken }) => (
-  <RootStackNav.Navigator headerMode="none">
-    { !!userToken ? (
-      <RootStackNav.Screen
-        name='App'
-        component={AppStack}
-      />
-    ) : (
-      <RootStackNav.Screen
-        name='Auth'
-        component={AuthStack}
-      />
-    )}
-  </RootStackNav.Navigator>
-)
+// const AppStack = () => {
+//   return (
+//     <AppStackNav.Navigator initialRouteName={'Home'}>
+//       <AppStackNav.Screen
+//         name="Home"
+//         component={HomePage}
+//         options={{
+//           title: '' ,
+//           // we chouldnt really be able to go back to the login page?
+//           // headerLeft: () => null
+//         }}
+//       />
+//       <AppStackNav.Screen
+//         name="Search"
+//         component={Search}
+//         options={{ title: '' }}
+//       />
+//       <AppStackNav.Screen
+//         name="Chat"
+//         component={Chat}
+//         options={{ title: 'Test Subject' }}
+//       />
+//       <AppStackNav.Screen
+//         name="ChatPicker"
+//         component={ChatPicker}
+//         options={{
+//           title: 'My Chats',
+//         }}
+//       />
+//       <AppStackNav.Screen
+//         name="MyProfile"
+//         component={Profile}
+//         options={{ title: 'My Profile' }}
+//       />
+//       <AppStackNav.Screen
+//         name="AdminPage"
+//         component={AdminPage}
+//         options={{ title: 'Admin Page' }}
+//       />
+//       <AppStackNav.Screen
+//         name='CreateChat'
+//         component={CreateChat}
+//         options={{ title: 'New Chat'}}
+//       />
+//       <AppStackNav.Screen
+//         name='Settings'
+//         component={Settings}
+//         options={{ title: 'Settings'}}
+//       />
+//     </AppStackNav.Navigator>
+//   )
+// }
+// const RootStackNav = createStackNavigator();
+// const RootStack = ({ userToken }) => (
+//   <RootStackNav.Navigator headerMode="none">
+//     { !!userToken ? (
+//       <RootStackNav.Screen
+//         name='App'
+//         component={AppStack}
+//       />
+//     ) : (
+//       <RootStackNav.Screen
+//         name='Auth'
+//         component={AuthStack}
+//       />
+//     )}
+//   </RootStackNav.Navigator>
+// )
 // const reducer = () =>
 
 const App = () => {
-  const [{authLoading, userToken, signingOut}, dispatch] = React.useReducer(
-    ( prevState, action) => {
-      switch (action.type) {
-        case 'CHECK_LOGIN':
-          return {
-            ...prevState,
-            authLoading: false,
-            userToken: action.token
-          };
-        case 'LOGIN':
-          return {
-            ...prevState,
-            loggingOut: false,
-            userToken: action.token
-          };
-        case 'LOGOUT':
-          return {
-            ...prevState,
-            loggingOut: true,
-            userToken: null
-          };
-      }
-    }, {
-      authLoading: true,
-      userToken: null,
-      loggingOut: false
-    }
-  )
+  // const [{authLoading, userToken, signingOut}, dispatch] = React.useReducer(
+  //   ( prevState, action) => {
+  //     switch (action.type) {
+  //       case 'CHECK_LOGIN':
+  //         return {
+  //           ...prevState,
+  //           authLoading: false,
+  //           userToken: action.token
+  //         };
+  //       case 'LOGIN':
+  //         return {
+  //           ...prevState,
+  //           loggingOut: false,
+  //           userToken: action.token
+  //         };
+  //       case 'LOGOUT':
+  //         return {
+  //           ...prevState,
+  //           loggingOut: true,
+  //           userToken: null
+  //         };
+  //     }
+  //   }, {
+  //     authLoading: true,
+  //     userToken: null,
+  //     loggingOut: false
+  //   }
+  // )
 
-  const _checkAuth = async () => {
-    let userToken;
-    try {
-      userToken = await AsyncStorage.getItem(LOGIN_TOKEN)
-    } catch (e) {
-      console.log('checking for login failed')
-    }
-    console.log('userToken', userToken)
-    dispatch({ type: 'CHECK_LOGIN', token: userToken})
-  }
+  // const _checkAuth = async () => {
+  //   let userToken;
+  //   try {
+  //     userToken = await AsyncStorage.getItem(LOGIN_TOKEN)
+  //   } catch (e) {
+  //     console.log('checking for login failed')
+  //   }
+  //   console.log('userToken', userToken)
+  //   dispatch({ type: 'CHECK_LOGIN', token: userToken})
+  // }
 
-  React.useEffect(() => {
-    _checkAuth()
-  }, [])
+  // React.useEffect(() => {
+  //   _checkAuth()
+  // }, [])
 
   const [user, setUser] = React.useState<IUser>({} as IUser);
   const updateUser = (newUser: IUser) => setUser(newUser)
@@ -218,66 +219,67 @@ const App = () => {
     setUser: updateUser
   }
 
-  const [_login, { data, error, loading}] = useMutation<ILoginPayload>(
-    LOGIN,
-    {
-      onCompleted: async ({ login }) => {
-        if (login.res) {
-          console.log('login result', login.user)
-          setUser({...login.user});
-          await AsyncStorage.setItem(LOGIN_TOKEN, '_loggedIn')
-          dispatch({ type: 'LOGIN', token: LOGIN_TOKEN})
-        } else {
+  // const [_login, { data, error, loading}] = useMutation<ILoginPayload>(
+  //   LOGIN,
+  //   {
+  //     onCompleted: async ({ login }) => {
+  //       if (login.res) {
+  //         console.log('login result', login.user)
+  //         setUser({...login.user});
+  //         await AsyncStorage.setItem(LOGIN_TOKEN, '_loggedIn')
+  //         dispatch({ type: 'LOGIN', token: LOGIN_TOKEN})
+  //       } else {
 
-        }
-      }
-    }
-  )
+  //       }
+  //     }
+  //   }
+  // )
 
 
-  const authContext = React.useMemo(
-    () => ({
-      login: (email: string, password: string) => {
-        // this is where we will call the mutation
-        console.log('data in login context', email, password)
-        _login({ variables: { email, password }})
+  // const authContext = React.useMemo(
+  //   () => ({
+  //     login: (email: string, password: string) => {
+  //       // this is where we will call the mutation
+  //       console.log('data in login context', email, password)
+  //       _login({ variables: { email, password }})
 
-        // dispatch({ type: 'LOGIN', token: LOGIN_TOKEN})
-      },
-      logout: async data => {
-        console.log('signing out in app')
-        await AsyncStorage.clear();
-        dispatch({ type: 'LOGOUT'})
-      },
-      createUser: data => {
-        dispatch({ type: 'LOGOUT', token: LOGIN_TOKEN})
-      },
-    }),
-    []
-  );
+  //       // dispatch({ type: 'LOGIN', token: LOGIN_TOKEN})
+  //     },
+  //     logout: async data => {
+  //       console.log('signing out in app')
+  //       await AsyncStorage.clear();
+  //       dispatch({ type: 'LOGOUT'})
+  //     },
+  //     createUser: data => {
+  //       dispatch({ type: 'LOGOUT', token: LOGIN_TOKEN})
+  //     },
+  //   }),
+  //   []
+  // );
 
-  if (authLoading) {
-    return (
-      <ActivityIndicator>
-        <View>
-          <Text>we are loading</Text>
-        </View>
-      </ActivityIndicator>
-    )
-  }
+  // if (authLoading) {
+  //   return (
+  //     <ActivityIndicator>
+  //       <View>
+  //         <Text>we are loading</Text>
+  //       </View>
+  //     </ActivityIndicator>
+  //   )
+  // }
 
   return (
-    <AuthContext.Provider value={authContext}>
-      <context.Provider value={value}>
-        <ThemeProvider theme={theme}>
-          <ApolloProvider client={client}>
+    <ApolloProvider client={client}>
+      <ThemeProvider theme={theme}>
+        <context.Provider value={value}>
+          {/* <AuthContext.Provider value={authContext}>
             <NavigationContainer>
               <RootStack userToken={userToken}/>
             </NavigationContainer>
-          </ApolloProvider>
-        </ThemeProvider>
-      </context.Provider>
-    </AuthContext.Provider>
+          </AuthContext.Provider> */}
+          <StackNavigation />
+        </context.Provider>
+      </ThemeProvider>
+    </ApolloProvider>
   )
 };
